map in level with type int

+++ src/de/kbgame/map/Level.java
@@ -6,7 +6,7 @@ import de.kbgame.util.ImageKey;
 public class Level {
 
 	public int width, height;
-	public byte[][] map;
+	public int[][] map;
 
 	private int left, right, top, bottom;
 
@@ -15,7 +15,7 @@ public class Level {
 	public Level(int wi, int hi) {
 		width = Math.max(10, wi);
 		height = Math.max(10, hi);
-		map = new byte[width][height];
+		map = new int[width][height];
 	}
 
 	public void update(Game g) {
@@ -31,7 +31,7 @@ public class Level {
 	public void draw(Game g) {
 		for (int x = left; x < right; x++) {
 			for (int y = top; y < bottom; y++) {
-				byte val = getMap(x, y);
+				int val = getMap(x, y);
 				switch (val) {
 				case Blocks.Solid: {
 					g.graphic.drawImage(ImageKey.BLOCK, x * BLOCK_WIDTH
@@ -45,13 +45,13 @@ public class Level {
 		}
 	}
 	
-	public byte getMap(int x, int y){
+	public int getMap(int x, int y){
 		if (x >= 0 && y >= 0 && x < width && y < height){
 			return map[x][y];
 		}else return 1;
 	}
 	
-	public void setMap(int x, int y, byte v){
+	public void setMap(int x, int y, int v){
 		if (x >= 0 && y >= 0 && x < width && y < height){
 			map[x][y] = v;
 		}
